BUG need to use conftest for fixtures
